+++ R/coor.test.R
@@ -1,7 +1,8 @@
 #' Randomization test for singing coordination 
 #' 
 #' Monte Carlo randomization test to assess the statistical significance of singing coordination
-#' @usage coor.test(X, iterations = 1000, less.than.chance = TRUE, parallel = 1, pb = TRUE)
+#' @usage coor.test(X, iterations = 1000, less.than.chance = TRUE, parallel = 1, pb = TRUE, 
+#' rm.imcomp = FALSE, cutoff = 2, rm.solo = FALSE)
 #' @param  X Data frame containing columns for singing event (sing.event), 
 #' individual (indiv), and start and end time of signal (start and end).
 #' @param iterations number of iterations for shuffling and calculation of the expected number of overlaps. Default is 1000.
@@ -12,6 +13,15 @@
 #'  It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).
 #' @param pb Logical argument to control progress bar. Default is \code{TRUE}. Note that progress bar is only used
 #' when parallel = 1.
+#' @param rm.imcomp Logical. If \code{TRUE} removes the events that don't have 2 interacting individuals. Default is
+#'  \code{FALSE}.
+#' @param cutoff Numeric. Determines the minimum number of signals per individual in a singing event. Events not meeting 
+#' this criterium are removed if rm.imcomp is \code{TRUE}. If rm.icomp is \code{FALSE} cutoff is ignored. Default is 2. 
+#' Note that randomization tests are not reliable with very small sample sizes. Ideally 10 or more signals per individual 
+#' should be available in each singing event.
+#' @param rm.solo Logical. Controls if signals that are not intercalated at the start or end of the 
+#' sequence are removed (if \code{TRUE}). For instances the sequence of signals A-A-A-B-A-B-A-B-B-B (in which A and B represent different individuals, as in the 'indiv' column) would be subset to 
+#' A-B-A-B-A-B. Default is  \code{FALSE}.
 #' @return A data frame with the observed number of overlaps (obs.overlaps), mean number of overlaps expected by chance,
 #' and p value.  
 #' @export
@@ -38,7 +48,8 @@
 #' @author Marcelo Araya-Salas (\email{araya-salas@@cornell.edu})
 #last modification on jul-5-2016 (MAS)
 
-coor.test <- function(X = NULL, iterations = 1000, less.than.chance = TRUE, parallel = 1, pb = TRUE)
+coor.test <- function(X = NULL, iterations = 1000, less.than.chance = TRUE, parallel = 1, pb = TRUE, 
+                      rm.imcomp = FALSE, cutoff = 2, rm.solo = FALSE)
 {
   if(!is.data.frame(X))  stop("X is not a data frame")
   
@@ -53,20 +64,30 @@ coor.test <- function(X = NULL, iterations = 1000, less.than.chance = TRUE, para
   #if there are NAs in start or end stop
   if(any(is.na(c(X$end, X$start)))) stop("NAs found in start and/or end")  
   
+  #remove hidden levels
+  X <- droplevels(X)
   
   #stop if some events do not have 2 individuals 
-      qw <- as.data.frame((tapply(X$sing.event, list(X$sing.event, X$indiv), length)))
+  qw <- as.data.frame(tapply(X$sing.event, list(X$sing.event, X$indiv), length))
+  qw[is.na(qw)] <- 0
+   
+   #complete singing events
+    if(rm.imcomp) cse <- qw[,1] >= cutoff & qw[,2] >= cutoff else cse <- qw[,1] >= 1 & qw[,2] >= 1
+    
+if(rm.imcomp)   {X <- X[X$sing.event %in% unique(X$sing.event)[cse], ]
+if(any(!cse)) warning("Some events didn't have 2 individuals and were excluded")
+} else
+  if(any(!cse)) stop("Some singing events don't have 2 interacting individuals ('indiv' colum)")
 
-   qw[qw > 0] <- 1
-
-if(any(apply(qw, 1, sum) != 2)) stop("Some singing events don't have 2 interating individuals ('indiv' colum)")
-
+    #if nothing was left
+    if(nrow(X) == 0) stop("No events left after removing incomplete events")
+    
   #if iterations is not vector or length==1 stop
   if(any(!is.vector(iterations),!is.numeric(iterations))) stop("'interations' must be a numeric vector of length 1") else{
     if(!length(iterations) == 1) stop("'interations' must be a numeric vector of length 1")}
   
    
-  interations <- round(iterations)
+  iterations <- round(iterations)
   
   #interations should be positive
   if(iterations < 1) stop("'iterations' must be a positive integer")
@@ -78,24 +99,26 @@ if(any(apply(qw, 1, sum) != 2)) stop("Some singing events don't have 2 interatin
     #create function to randomized singing events
       coortestFUN <- function(X, h){
   sub<-X[X$sing.event==h,]
-  
-  #remove solo singing
   sub<-sub[order(sub$start),]
   
-  fst <- max(c(which(sub$start==min(sub$start[sub$indiv==unique(sub$indiv)[1]])),
+  #remove solo singing
+  if(rm.solo)
+    {  
+    fst <- max(c(which(sub$start==min(sub$start[sub$indiv==unique(sub$indiv)[1]])),
                which(sub$start==min(sub$start[sub$indiv==unique(sub$indiv)[2]])))) - 1
   
   lst <- min(c(which(sub$start==max(sub$start[sub$indiv==unique(sub$indiv)[1]])),
                which(sub$start==max(sub$start[sub$indiv==unique(sub$indiv)[2]])))) + 1
   
   sub <- sub[fst:lst, ]
+  }
   
   sub1<-sub[sub$indiv==unique(sub$indiv)[1], ]
   sub2<-sub[sub$indiv==unique(sub$indiv)[2], ]
   
   
   #determine which ones overlap (observed)
-  ovlp<-sapply(2:nrow(sub),function(i) {
+  ovlp <- sapply(2:nrow(sub),function(i) {
     if(sub$start[i]>sub$start[i-1] & sub$start[i]<sub$end[i-1])  
       "ovlp" else "No ovlp"})