+++ R/imp.syrinx.R
@@ -2,12 +2,13 @@
 #' 
 #' \code{imp.syrinx} imports Syrinx selection data from many files simultaneously. 
 #' All files must be have the same columns.
-#' @usage imp.syrinx(path = NULL, all.data = FALSE, recursive = FALSE)  
+#' @usage imp.syrinx(path = NULL, all.data = FALSE, recursive = FALSE, exclude = FALSE)  
 #' @param path A character string indicating the path of the directory in which to look for the text files. 
 #' If not provided (default) the function searches into the current working directory. Default is \code{NULL}).
 #' @param all.data Logical. If \code{TRUE}) all columns in text files are returned. Default is \code{FALSE}). Note 
 #' that all files should contain exactly the same columns in the same order. 
 #' @param recursive Logical. If \code{TRUE}) the listing recurse into sub-directories.
+#' @param exclude Logical. Controls whether files that cannot be read are ignored (\code{TRUE}). Default is \code{FALSE}.
 #' @return A single data frame with information of the selection files. If all.data argument is set to \code{FALSE}) the data 
 #' frame contains the following columns: selec, start, end, and selec.file. If sound.file.col is provided the data frame
 #' will also contain a 'sound.files' column. In addition, all rows with duplicated data are removed. This is useful when 
@@ -42,7 +43,7 @@
 #' @author Marcelo Araya-Salas (\email{araya-salas@@cornell.edu})
 #last modification on jul-5-2016 (MAS)
 
-imp.syrinx <- function(path = NULL, all.data = FALSE, recursive = FALSE) 
+imp.syrinx <- function(path = NULL, all.data = FALSE, recursive = FALSE, exclude = FALSE) 
 { 
   
   #check path to working directory
@@ -65,28 +66,36 @@ if(substring(text = readLines(sel.txt[1])[1], first = 0, last = 9) == "fieldkey:
 
 clist<-lapply(1:length(sel.txt), function(i)
   {    
-  if(field)  {a <- read.table(sel.txt[i], header = TRUE, sep = "\t", fill = TRUE) 
-
-  if(!all.data) { c <- data.frame(selec.file = sel.txt2[i], sound.files = a[, grep("soundfile",colnames(a))],
+  if(field)  {
+    
+    a <- try(read.table(sel.txt[i], header = TRUE, sep = "\t", fill = TRUE), silent = TRUE) 
+    if(!exclude & class(a) == "try-error") stop(paste("The selection file",sel.txt[i], "cannot be read"))
+    
+  if(!class(a) == "try-error" & !all.data) { c <- data.frame(selec.file = sel.txt2[i], sound.files = a[, grep("soundfile",colnames(a))],
                                 selec = 1,
                                 start = a[,grep("lefttimesec",colnames(a))],
                                 end = a[,grep("righttimesec",colnames(a))])
   for(i in 2:nrow(c)) if(c$selec.file[i] == c$selec.file[i-1]) c$selec[i]<-c$selec[i-1] + 1
   } else c<-a 
-                                b<-rbind(b, c)
                                 } else {
-          a<-read.table(sel.txt[i],header = FALSE,sep = "\t", fill = TRUE)
-            c <- a[, seq(2, ncol(a), by =2)]
+            a <- try(read.table(sel.txt[i], header = FALSE, sep = "\t", fill = TRUE), silent = TRUE) 
+            if(!exclude & class(a) == "try-error") stop(paste("The selection file",sel.txt[i], "cannot be read"))
+            
+            if(!class(a) == "try-error") 
+              { 
+              c <- a[, seq(2, ncol(a), by =2)]
            colnames(c) <- gsub(":", "", unlist(a[1, seq(1,ncol(a), by =2)]), fixed = TRUE)
            if(!all.data) {c<-data.frame(sound.files = c[, grep("selected",colnames(c), ignore.case = TRUE)],
                                        selec = 1,
                                        start = c[, grep("lefttime",colnames(c), ignore.case = TRUE)],
                                        end = c[, grep("righttime",colnames(c), ignore.case = TRUE)])
-           for(i in 2:nrow(c)) if(c$sound.files[i] == c$sound.files[i-1]) c$selec[i] <- c$selec[i-1] + 1}          
+           for(i in 2:nrow(c)) if(c$sound.files[i] == c$sound.files[i-1]) c$selec[i] <- c$selec[i-1] + 1} 
+           } else c <- a         
+                                }
   return(c)
-                    }
 })
 
+clist <- clist[sapply(clist, is.data.frame)]
 b <- do.call("rbind", clist)
 if(!all.data) if(any(is.na(b$start))) warning("NAs found (empty rows)")
 return(b[!duplicated(b), ])